+++ Runtime/Core/IConsolePilotOpenCloseTarget.cs
@@ -1,0 +1,15 @@
+namespace ConsolePilot
+{
+    public interface IConsolePilotOpenCloseTarget
+    {
+        bool IsOpen { get; }
+
+        void SetOpen(bool isOpen);
+
+        void Open();
+
+        void Close();
+
+        void Toggle();
+    }
+}